+++ src/addons/display_commander/hooks/loadlibrary_hooks.cpp
@@ -20,7 +20,6 @@
 #include "../utils/detour_call_tracker.hpp"
 #include "../utils/general_utils.hpp"  // GetDefaultDlssOverrideFolder, GetCallingDLL
 #include "../utils/logging.hpp"
-#include "../utils/pe_static_imports.hpp"
 #include "../utils/timing.hpp"
 #include "ddraw/ddraw_present_hooks.hpp"
 #include "hook_suppression_manager.hpp"
@@ -1529,15 +1528,9 @@ void OnModuleLoaded(const std::wstring& moduleName, HMODULE hModule) {
             }
         }
     }
-    const std::wstring module_display_name = (IsKnownDllName(moduleName) ? L"*" : L"") + moduleName;
+    const std::wstring module_display_name = moduleName;
     LogInfo("[OnModuleLoaded] %ws (0x%p)%s%s", module_display_name.c_str(), hModule, has_dc_export ? " (DC proxy)" : "",
             product_and_version.c_str());
-    {
-        std::string imports = GetStaticImportDllNamesSingleLine(hModule);
-        if (!imports.empty()) {
-            LogInfo("[OnModuleLoaded] %ws static imports: %s", module_display_name.c_str(), imports.c_str());
-        }
-    }
 
     if (IsRenoDxAddonPath(moduleName)) {
         OnRenoDxAddonLoaded();